+++ CHANGELOG.md
@@ -1,4 +1,20 @@
 # 更新日志\n
+## [ForwardWidgets-huangxd] - 2025-12-31 03:18:12
+- 仓库: https://github.com/huangxd-/ForwardWidgets
+- 分支: main
+- 新增: ad2a81ae257e95bdd76b458173f52e9ee488020e
+
+## [ForwardWidgets-2kuai] - 2025-12-31 03:18:12
+- 仓库: https://github.com/2kuai/ForwardWidgets
+- 分支: main
+- 新增: 58dfeac184f9d70f78934fa982bfd5a4b2789353
+
+## [ForwardWidgets-pack1r] - 2025-12-31 03:18:12
+- 仓库: https://github.com/pack1r/ForwardWidgets
+- 分支: main
+- 新增: 0adb054f9b64f46661d13e0c7d2291b4d3e02db0
+
+
 ## [ForwardWidgets-huangxd] - 2025-12-30 03:19:00
 - 仓库: https://github.com/huangxd-/ForwardWidgets
 - 分支: main